avances niveles de inversor

+++ src/main/java/com/systeam/investment/dto/InvestmentResponse.java
@@ -19,7 +19,7 @@ public class InvestmentResponse {
     private BigDecimal precioSubtoken;
     private String txHash;
     private String estado;
-    private Integer descuentoPorcentaje;
+    private BigDecimal descuentoPorcentaje;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
